requirements.txt
# Gerekli temel kütüphaneler (örnek)
speechrecognition
pyttsx3
llama-cpp-python
